Correct dependencies and configuration after Artemis switch
Unlike ActiveMQ, Artemis does not pull in Jackson transitively so
a dependency on spring-boot-starter-json is required to be able to
use MappingJackson2MessageConverter.

To avoid having to run an external broker, the guide uses an embedded
broker. However, Artemis does not run in embedded mode by default and
spring-boot-starter-artemis does not contain the necessary
dependencies. Configuration properties and a dependency on
artemis-jakarta-server have been added to allow Artemis to run in
embedded mode.